status-receiver/requirements.txt
Flask==2.1.1
kafka-python==2.0.2